Update previous tests and add tests for auth
